Fixing the code to match the tests

+++ myBank.h
@@ -4,7 +4,7 @@ void CheckingBalance(int);
 void Deposit(int,double);
 void Withdraw(int, double);
 void CloseBankAccount(int);
-void AddInterest(double);
+void AddInterest(int);
 void PrintingAccounts();
 void ClosingAllAccounts();
 double decimal(double);
